refactor wave_vision structure to separate detectors and descriptors
