+++ wave_vision/CMakeLists.txt
@@ -23,14 +23,15 @@ ADD_LIBRARY(
     wave_vision
     STATIC
     src/common.cpp
-    src/fast_detector.cpp
+    src/detector/fast_detector.cpp
 )
 TARGET_LINK_LIBRARIES(wave_vision ${WAVE_VISION_DEPS})
 
 # UNIT TESTS
-WAVE_ADD_TEST(${PROJECT_NAME}_tests
-              tests/common_tests.cpp
-              tests/fast_tests.cpp)
+WAVE_ADD_TEST(
+    ${PROJECT_NAME}_tests
+    tests/common_tests.cpp
+    tests/detector_tests/fast_tests.cpp)
 TARGET_LINK_LIBRARIES(
     ${PROJECT_NAME}_tests
     ${PROJECT_NAME}
