+++ wave_vision/config/detector/fast.yaml
