+++ wave_vision/include/wave/vision/detector/fast_detector.hpp
@@ -12,7 +12,7 @@
 #include <vector>
 
 /** Libwave Headers */
-#include "wave/vision/feature_detector.hpp"
+#include "feature_detector.hpp"
 
 /** The wave namespace */
 namespace wave {
