+++ wave_vision/src/detector/fast_detector.cpp
@@ -1,5 +1,5 @@
 // Libwave Headers
-#include "wave/vision/fast_detector.hpp"
+#include "wave/vision/detector/fast_detector.hpp"
 
 namespace wave {
