+++ wave_vision/tests/config/detector/fast.yaml
